# cloudpdf/admin-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CloudPDF Admin API",
    "version": "3.0.0-next.0",
    "description": "Tenant-scoped admin surface of the CloudPDF document engine. Generated from the @cloudpdf/admin-api operation registry — do not edit by hand.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    }
  },
  "paths": {
...
    "/v1/tenants/{tenantId}/tokens": {
      "post": {
        "operationId": "tokens.issue",
        "summary": "Mint a delegated JWT: a doc token, or (API token only) a tenant token.",
        "description": "kind \"tenant\" requires the API token — authority mints only downward. Mounted only when the deployment can sign (HS256 mode); asymmetric deployments mint with their own private key.",
        "security": [
          {
            "apiToken": []
          },
          {
            "tenantToken": []
          }
        ],
        "x-required-scope": [
          "tokens.issue-doc"
        ],
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]+$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "type": "object",
                    "properties": {
                      "kind": {
                        "type": "string",
                        "enum": [
                          "doc"
                        ]
                      },
                      "sub": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 256
                      },
                      "docId": {
                        "type": "string",
                        "pattern": "^[A-Za-z0-9_-]+$"
                      },
                      "layerName": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 256
                      },
                      "scope": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 128
                        },
                        "minItems": 1,
                        "maxItems": 64
                      },
                      "userId": {
                        "type": "string",
                        "maxLength": 256
                      },
                      "displayName": {
                        "type": "string",
                        "maxLength": 256
                      },
                      "groupId": {
                        "type": "string",
                        "maxLength": 256
                      },
                      "groups": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "maxLength": 256
                        },
                        "maxItems": 64
                      },
                      "expiresIn": {
                        "type": "integer",
                        "minimum": 60,
                        "maximum": 7776000
                      }
                    },
                    "required": [
                      "kind",
                      "sub",
                      "docId",
                      "scope",
                      "expiresIn"
                    ],
                    "additionalProperties": false
                  },
                  {
                    "type": "object",
                    "properties": {
                      "kind": {
                        "type": "string",
                        "enum": [
                          "tenant"
                        ]
                      },
                      "sub": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 256
                      },
                      "scope": {
                        "type": "array",
                        "items": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "*"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "docs.create",
                                "docs.read",
                                "docs.delete",
                                "tokens.issue-doc",
                                "tokens.revoke"
                              ]
                            }
                          ]
                        },
                        "minItems": 1,
                        "maxItems": 16
                      },
                      "expiresIn": {
                        "type": "integer",
                        "minimum": 60,
                        "maximum": 7776000
                      }
                    },
                    "required": [
                      "kind",
                      "sub",
                      "scope",
                      "expiresIn"
                    ],
                    "additionalProperties": false
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    },
                    "jti": {
                      "type": "string"
                    },
                    "expiresAt": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "token",
                    "jti",
                    "expiresAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorPayload"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "The deployment's static root credential (CLOUDPDF_API_AUTH_TOKENS), valid on every surface."
      },
      "tenantToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Delegated tenant JWT, valid only under its own /v1/tenants/{tenantId}/ subtree — the path tenant must equal the token's tenant_id. Doc-scoped viewer tokens are rejected on every admin route."
      }
    },
    "schemas": {
      "AdminErrorPayload": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      }
    }
  }
}